+++ CHANGES.md
@@ -29,6 +29,8 @@ Changes in 12.7
 - Use MultipleCheckboxInput() function in `_makeMultipleInput()` in StudentsUsersInfo.fnc.php
 - Fix PHP8.5 deprecated curl_close() function in curl.php
 - Double (free) Comment input size when MP does not do Comments in InputFinalGrades.php
+- Misconfiguration hint: Scale Value different from max GPA Value in ReportCardGrades.php
+- Misconfiguration hint: Minimum Passing Grade > Scale Value in ReportCardGrades.php
 
 Changes in 12.6
 ---------------